Updated multiple back-end functions to:
- Fix  unknown error handling by casting errors to Error where needed
- Add explicit typing for several array/map callbacks in fetch-daily-focus
- Casts for infographic encoding result
- Migrate push notification sender from web-push to native Web Crypto implementation (rewritten send-push-notification)
- Clean up test and trust score flows accordingly
- Remove reliance on web-push library and replace with VAPID JWT signing and AES-GCM encryption
This includes SQL/migration edits and TS updates across cleanup-expired-cache, fetch-daily-focus, generate-infographic, get-trust-score, and test-supadata, plus the new native Web Push path in send-push-notification.

X-Lovable-Edit-ID: edt-e7f29001-dc0c-41ce-8a0c-8f1c9de67f6d

+++ supabase/functions/generate-infographic/index.ts
@@ -166,7 +166,7 @@ ${truncatedText}`;
       // Composite the background (with logo) onto the infographic
       (infograph as any).composite(logoBackground, x, y);
 
-      finalBuffer = await infograph.encode();
+      finalBuffer = await infograph.encode() as Uint8Array;
       watermarkStatus = 'success (with background)';
       console.log('[Infographic] Watermark applied with background');
     } catch (err: any) {
